Revert use of cuboid
