+++ docs/guides/quickfort-library-guide.rst
@@ -45,7 +45,9 @@ and a convenient `checklist
 <https://docs.google.com/spreadsheets/d/15TDBebP8rBNvsFbezb9xuKPmGWNzv7j4XZWq1AsfCio/edit#gid=1459509569>`__
 that you can use to track your progress.
 
-If you like, you can download a fully built Dreamfort-based fort from
+If you'd like a visual demonstration, there is a `series of videos on YouTube <https://www.youtube.com/playlist?list=PLzXx9JcB9oXxmrtkO1y8ZXzBCFEZrKxve>`__
+that walk you through the entire process of building a Dreamfort-based
+fortress. You can also download a fully built Dreamfort-based fort from
 :dffd:`dffd <15434>`, load it, and explore it interactively.
 
 Here are annotated screenshots of the major Dreamfort levels (or click `here
