+++ docs/plugins/dwarfvet.rst
@@ -16,6 +16,8 @@ unassigned from the pasture while in treatment. The animal will be reassigned
 to its original pasture shortly after recovery.
 
 Animals that are on restraints or in cages will not be designated for treatment.
+Animals require an empty tile of floor to rest on while recovering, so your
+hospital must have some unoccupied floor space in order for animals to use it.
 
 You can enable ``dwarfvet`` in `gui/control-panel`, and you can choose to start
 ``dwarfvet`` automatically in new forts in the ``Gameplay`` -> ``Autostart``
