+++ library/Core.cpp
@@ -93,6 +93,7 @@ using namespace DFHack;
 using namespace df::enums;
 using df::global::init;
 using df::global::world;
+using std::string;
 
 // FIXME: A lot of code in one file, all doing different things... there's something fishy about it.
 
@@ -1492,7 +1493,7 @@ Core::Core() :
     color_ostream::log_errors_to_stderr = true;
 };
 
-void Core::fatal (std::string output)
+void Core::fatal (std::string output, const char * title)
 {
     errorstate = true;
     std::stringstream out;
@@ -1509,7 +1510,9 @@ void Core::fatal (std::string output)
     fprintf(stderr, "%s\n", out.str().c_str());
     out << "Check file stderr.log for details.\n";
     std::cout << "DFHack fatal error: " << out.str() << std::endl;
-    DFSDL::DFSDL_ShowSimpleMessageBox(0x10 /* SDL_MESSAGEBOX_ERROR */, "DFHack error!", out.str().c_str(), NULL);
+    if (!title)
+        title = "DFHack error!";
+    DFSDL::DFSDL_ShowSimpleMessageBox(0x10 /* SDL_MESSAGEBOX_ERROR */, title, out.str().c_str(), NULL);
 
     bool is_headless = bool(getenv("DFHACK_HEADLESS"));
     if (is_headless)
@@ -1611,7 +1614,27 @@ bool Core::InitMainThread() {
         }
         else
         {
-            fatal("Not a known DF version.\n");
+            std::stringstream msg;
+            msg << "Not a supported DF version.\n"
+                   "\n"
+                   "Please make sure that you have a version of DFHack installed that\n"
+                   "matches the version of Dwarf Fortress.\n"
+                   "\n"
+                   "DFHack version: " << Version::dfhack_version() << "\n"
+                   "\n";
+            auto supported_versions = vif->getVersionInfosForCurOs();
+            if (supported_versions.size()) {
+                msg << "Dwarf Fortress releases supported by this version of DFHack:\n\n";
+                for (auto & sv : supported_versions) {
+                    string ver = sv->getVersion();
+                    if (ver.starts_with("v0.")) {  // translate "v0.50" to the standard format: "v50"
+                        ver = "v" + ver.substr(3);
+                    }
+                    msg << "    " << ver << "\n";
+                }
+                msg << "\n";
+            }
+            fatal(msg.str(), "DFHack version mismatch");
         }
         errorstate = true;
         return false;
