+++ library/include/VersionInfo.h
@@ -75,6 +75,17 @@ namespace DFHack
             OS = rhs.OS;
         };
 
+        static OSType getCurOS() {
+#if defined(_WIN32)
+            const OSType expected = OS_WINDOWS;
+#elif defined(_DARWIN)
+            const OSType expected = OS_APPLE;
+#else
+            const OSType expected = OS_LINUX;
+#endif
+            return expected;
+        }
+
         uintptr_t getBase () const { return base; };
         intptr_t getRebaseDelta() const { return rebase_delta; }
         void setBase (const uintptr_t _base) { base = _base; };
@@ -171,13 +182,8 @@ namespace DFHack
         };
 
         void ValidateOS() {
-#if defined(_WIN32)
-            const OSType expected = OS_WINDOWS;
-#elif defined(_DARWIN)
-            const OSType expected = OS_APPLE;
-#else
-            const OSType expected = OS_LINUX;
-#endif
+            static const OSType expected = getCurOS();
+
             if (expected != getOS()) {
                 std::cerr << "OS mismatch; resetting to " << int(expected) << std::endl;
                 setOS(expected);
